formatting tweaks + sablier added to wallet balance
