+++ src/utils/format.js
@@ -31,6 +31,26 @@ import { getCurrencySymbol } from 'utils/common';
  * - accepts null and undefined values but returns null instead of formatted value as well as NaNs and ininity,
  */
 
+type FormatValueOptions = {|
+  decimalPlaces?: number; // default: undefined -> truncate factional zeros
+|};
+
+/**
+ * 1000 => 1,000
+ */
+export function formatValue(value: ?BigNumber, options?: FormatValueOptions) {
+  if (!value || !value.isFinite()) return null;
+
+  const format = {
+    decimalSeparator: '.',
+    groupSeparator: ',',
+    groupSize: 3,
+  };
+
+  // $FlowFixMe: incorrect flow-typed `decimalPlaces` nullability
+  return value.toFormat(options?.decimalPlaces, BigNumber.ROUND_DOWN, format);
+}
+
 /**
  * 0.5 => '50%'
  * -0.01234 => '-1,23%'
@@ -38,7 +58,7 @@ import { getCurrencySymbol } from 'utils/common';
 export function formatPercentValue(value: ?BigNumber) {
   if (!value || !value.isFinite()) return null;
 
-  return t('percentValue', { value: value.multipliedBy(100).precision(2).toFixed() });
+  return t('percentValue', { value: formatValue(value.multipliedBy(100), { decimalPlaces: 2 }) });
 }
 
 /**
@@ -49,15 +69,8 @@ export function formatPercentChange(value: ?BigNumber) {
   if (!value || !value.isFinite()) return null;
 
   return value.gte(0)
-    ? t('positivePercentValue', { value: value.multipliedBy(100).precision(2).toFixed() })
-    : t('percentValue', { value: value.multipliedBy(100).precision(2).toFixed() });
-}
-
-export function formatFiatValue(value: ?BigNumber, currency?: string) {
-  const formattedValue = formatValueWithUnit(value);
-  if (!formattedValue) return null;
-
-  return currency ? t('fiatValue', { value: formattedValue, symbol: getCurrencySymbol(currency) }) : formattedValue;
+    ? t('positivePercentValue', { value: formatValue(value.multipliedBy(100), { decimalPlaces: 2 }) })
+    : t('percentValue', { value: formatValue(value.multipliedBy(100), { decimalPlaces: 2 }) });
 }
 
 /**
@@ -74,7 +87,18 @@ export function formatValueWithUnit(value: ?BigNumber) {
   if (value.gte(threshold * 1e6)) return t('units.1e6', { value: value.dividedBy(1e6).toFixed(2) });
   if (value.gte(threshold * 1e3)) return t('units.1e3', { value: value.dividedBy(1e3).toFixed(2) });
   return value.toFixed(2);
-};
+}
+
+type FormatFiatValueOptions = {|
+  exact: boolean, // default: false
+|};
+
+export function formatFiatValue(value: ?BigNumber, currency?: string, options?: FormatFiatValueOptions) {
+  const formattedValue = options?.exact ? formatValue(value, { decimalPlaces: 2 }) : formatValueWithUnit(value);
+  if (!formattedValue) return null;
+
+  return currency ? t('fiatValue', { value: formattedValue, symbol: getCurrencySymbol(currency) }) : formattedValue;
+}
 
 /**
  * Formats profit as `+10.00% ($100.00)`.
